insights: move test
