+++ enterprise/internal/insights/background/queryrunner/testdata/TestQueryExecution_ToQueueJob/test_to_job_with_dependents.golden
@@ -1,0 +1,11 @@
+&queryrunner.Job{
+	SeriesID: "series1", SearchQuery: "sourcegraphquery1",
+	RecordTime: &time.Time{
+		ext: 63713433600,
+	},
+	Cost:            500,
+	Priority:        1000,
+	PersistMode:     "record",
+	DependentFrames: []time.Time{{ext: 63713520000}},
+	State:           "queued",
+}
